Add EFI hello world example
Need to split stuff into generic/linux/efi at some point

+++ hc/libc.h
@@ -7,8 +7,8 @@ typedef unsigned short uint16_t;
 typedef short int16_t;
 typedef unsigned int uint32_t;
 typedef int int32_t;
-typedef unsigned long uint64_t;
-typedef long int64_t;
+typedef unsigned long long uint64_t;
+typedef long long int64_t;
 
 #define INT8_MIN (-1 - 0x7f)
 #define INT16_MIN (-1 - 0x7fff)
